rename: 초기 데이터 초기화 클래스 패키지 변경
